+++ src/preview.h
@@ -9,6 +9,10 @@ gboolean preview_is_visible(NotesWindow *win);
 void     preview_queue_update(NotesWindow *win);
 void     preview_apply_theme(NotesWindow *win);
 void     preview_apply_layout(NotesWindow *win);
+void     preview_apply_font_size(NotesWindow *win);
+void     preview_find_show(NotesWindow *win);
+void     preview_find_hide(NotesWindow *win);
+void     preview_export_pdf(NotesWindow *win, const char *out_path);
 void     preview_zoom_step(NotesWindow *win, double delta);
 
 #endif